pyproject.toml
[project]
name = "q3webApi"
description = "Read q3 events from kafka and visualise them"
version = "0.0.1"
authors = [
    { name = "Marco Siebecke", email = "marco@siebecke.se" }
]
dependencies = [
    "click",
    "Flask",
    "Flask-SocketIO",
    "kafka-python",
    "gevent",
    "gevent-websocket"
]

[project.scripts]
q3webApi = "q3webApi.main:cli"

[tool.setuptools.packages.find]
exclude = ["q3web"]
